Only add schedule buttons when they are not there
This is a kind of workaround to the problem that the method
toggleSecondaryButtons is called more times than is should be.
If a schedule is added from the database which has the mode set to new
and updated apps or custom list and the spinner is then selected the
toggling method might be triggered more than once. This issue might only
happen in tests.