Clean database after each test instead of before
This way each test will start with a clean database.
Since the activity is started before the before-annotated methods,
cleaning the database in a @Before method made the activity start with
data from the previous test which was then immediately cleaned. This
dirty initial state caused some tests to fail.